Added some functionality
Ok this commit contains some new changes, I've added some classes and new variables to the baseappbar.dart file that keeps track of values selected from the start screen, I've made a new file that contains a list of classes with the ability to take an image, have a name, and type attribute to it. (this is great for when we have photos to add). Added a results page and counters respectivley that go to the next grid after a choice has been selected.

WHAT WE STILL NEED:
-Images to randomly be placed in our grid systems with a correct answer being chosen.
-the grid that loads up from gridview.count() needs to not be scrollable and has to show all clickable cards on the same screen
-32 default images must be made in a folder for us to pick from and test (not urgent yet)_