Custom Images
Custom Images is complete, there is a feature where it doesn't allow duplicate images under the same name, but it allows complete freedom of uploading any image from the gallery of the device! Also changed the results screen to better display what you picked wrong and what was correct. We are literally at the finish line.